Product Not Complete

+++ src/main/java/com/chanthai/phoneshop/exception/ResourceNotFoundException.java
@@ -3,7 +3,7 @@ package com.chanthai.phoneshop.exception;
 import org.springframework.http.HttpStatus;
 
 public class ResourceNotFoundException extends APIException{
-    public ResourceNotFoundException(String resourceName,Integer id) {
+    public ResourceNotFoundException(String resourceName,Long id) {
         super(HttpStatus.NOT_FOUND,String.format("%s With id = %d Not Found",resourceName,id));;
     }
 }
