+++ src/main/java/com/chanthai/phoneshop/service/ModelService.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface ModelService {
     Model save (Model model);
-    List<Model> getByBrand(Integer brandId);
+    List<Model> getByBrand(Long brandId);
 }
